% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsp_de_analysis_parallel.R
\name{check_input}
\alias{check_input}
\title{Internal helper function for checking input paratmeters}
\usage{
check_input(
  de_results,
  norm_counts,
  samp_notes,
  grouping_var,
  base_level,
  control_var,
  show_pb,
  min_sample_size
)
}
\description{
Internal helper function for checking input paratmeters
}
