% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_target_notes.R
\name{convert_target_notes}
\alias{convert_target_notes}
\title{convert_target_notes

Converts target notes from Erin to DSP-DA format}
\usage{
convert_target_notes(targetNotes)
}
\arguments{
\item{targetNotes}{target notes data frame}
}
\value{
targetNotes data frame formatted like DSP-DA
}
\description{
convert_target_notes

Converts target notes from Erin to DSP-DA format
}
\examples{
targetNotes <- convert_target_notes(targetNotes)

}
