% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_analysis.R
\name{de_genes_by_group}
\alias{de_genes_by_group}
\title{de_genes_by_group

Function used to determine unique de genes by de group}
\usage{
de_genes_by_group(
  de = de_results,
  pval = pval_cutoff,
  fc = fc_cutoff,
  group = grouping_var,
  SP = dfs[[3]]
)
}
\arguments{
\item{de}{de results dataframe defaults to de_results}

\item{pval}{pval cutoff defaults to pval_cutoff}

\item{fc}{fc cutoff defaults to fc_cutoff}

\item{group}{grouping variable defaults to grouping_var}

\item{SP}{Segment Properties defaults to dfs[[3]]}
}
\value{
list of de genes unique to each de group
}
\description{
de_genes_by_group

Function used to determine unique de genes by de group
}
\examples{

de_genes_by_group(de = de_results, pval = pval_cutoff, fc = fc_cutoff, group = grouping_var, SP = dfs[[3]])

}
