% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsp_de_analysis_parallel.R
\name{de_internal}
\alias{de_internal}
\title{Internal lower-level function that actually computes the DE.
returns the data.frame of results.
This function does the following:
1. Transposes and transforms the entire norm_counts dataset.
2. Create all pairwise combinations that have at least min_sample_size observations per group.
3. lapply over each element of #2.
4. Process over each gene in #3.
5. Collate data.frames (or lists) and return.}
\usage{
de_internal(
  norm_counts,
  samp_notes,
  grouping_var,
  base_level,
  control_var,
  n_processors,
  the_formula,
  show_pb,
  min_sample_size
)
}
\description{
Internal lower-level function that actually computes the DE.
returns the data.frame of results.
This function does the following:
1. Transposes and transforms the entire norm_counts dataset.
2. Create all pairwise combinations that have at least min_sample_size observations per group.
3. lapply over each element of #2.
4. Process over each gene in #3.
5. Collate data.frames (or lists) and return.
}
