% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsp_de_analysis_parallel.R
\name{de_logic}
\alias{de_logic}
\title{Internal lower-level function that checks de_ressults attributes and runs de_internal
if needed. Returns the de_results (whether it was computed, referenced by file, or referenced
by object).}
\usage{
de_logic(
  de_results,
  norm_counts,
  samp_notes,
  grouping_var,
  base_level,
  control_var,
  n_processors,
  the_formula,
  show_pb,
  min_sample_size
)
}
\description{
Internal lower-level function that checks de_ressults attributes and runs de_internal
if needed. Returns the de_results (whether it was computed, referenced by file, or referenced
by object).
}
