% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dataset.R
\name{drop_samples}
\alias{drop_samples}
\title{Function that drops samples that meet a given criteria}
\usage{
drop_samples(df, cols, conditions, operators, keep = FALSE, and = TRUE, custom)
}
\arguments{
\item{df}{list of data.frame objects}

\item{cols}{what column(s) can criteria be found in Segment Properties}

\item{conditions}{criteria to find in segment properties}

\item{keep}{should samples that match conditions be kept}

\item{and}{if multiple conditions, should samples match both conditions or just one}

\item{custom}{custom filtering equation}
}
\value{
list of data.frame objects with samples dropped
}
\description{
Function that drops samples that meet a given criteria
}
\examples{
drop_samples(cols = "tissue", conditions = "CRC7", keep = TRUE, and = TRUE, custom = NULL)
}
