% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_GSEA.R
\name{entrez_to_symbol}
\alias{entrez_to_symbol}
\title{entrez_to_symbol

Function to change entrez ID to gene symbol}
\usage{
entrez_to_symbol(X, sp = species)
}
\arguments{
\item{X}{GSEA dataframe with leading edge data}

\item{sp}{name of the species for which analysis applies, supports "human" or "mouse", defaults to species in config}
}
\value{
gene symbols
}
\description{
entrez_to_symbol

Function to change entrez ID to gene symbol
}
\examples{

entrez_to_symbol(X = fgseaRes, sp = species)

}
