% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_topGS.R
\name{find_topGS}
\alias{find_topGS}
\title{find_topGS

Generic call to heatmap functionality}
\usage{
find_topGS(
  targetNotes = NULL,
  de_results = NULL,
  name = NULL,
  outdir = NULL,
  fileType = NULL,
  width = 1500,
  height = 850
)
}
\arguments{
\item{targetNotes}{target annotations - list of lists of genes output from parse_GeneSets}

\item{de_results}{table of DE results}

\item{name}{annotations to include on heatmap}

\item{outdir}{where to print file}

\item{fileType}{function to call for file printing}

\item{width}{file width}

\item{height}{file height}
}
\value{

}
\description{
find_topGS

Generic call to heatmap functionality
}
\examples{


}
