% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSP_utils.R
\name{make_name_valid}
\alias{make_name_valid}
\title{function to coerce DSP-DA sample names to a valid format}
\usage{
make_name_valid(name_test)
}
\arguments{
\item{name_test}{string indicating sample identification}
}
\value{
string with \code{name_test} contained within backquote
}
\description{
function to coerce DSP-DA sample names to a valid format
}
\examples{
make_name_valid("CPA 1 | Slide 1 | Seg-A")
}
