% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSP_utils.R
\name{nanoNormFactors}
\alias{nanoNormFactors}
\title{function to calculate normalization factors based on geometric mean}
\usage{
nanoNormFactors(v)
}
\arguments{
\item{v}{numeric vector}
}
\value{
vector of nnumeric normalization factors for \code{v}
}
\description{
function to calculate normalization factors based on geometric mean
}
\examples{
nanoNormFactors(c(0, 1, 5, 8, 3))
}
