% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSP_utils.R
\name{ngeoMean}
\alias{ngeoMean}
\title{function to calculate the geometric mean of a vector}
\usage{
ngeoMean(v)
}
\arguments{
\item{v}{numeric vector}
}
\value{
geometric mean of \code{v}
}
\description{
function to calculate the geometric mean of a vector
}
\examples{
ngeoMean(c(0, 1, 5, 8, 3))
}
