% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dataset.R
\name{parallel_read}
\alias{parallel_read}
\title{Function that parallelizes spreadsheet read}
\usage{
parallel_read(file, get_cols = TRUE)
}
\arguments{
\item{file}{string value with file name}

\item{get_cols}{boolean indicating if column names should be preserved}
}
\value{
list of data.frame objects containing data from /code{file} tabs
}
\description{
Reads in DSP dataset
}
\examples{
parallel_read(file="./testData/DA_data/Q3_normalization/Q3 norm.xlsx")
}
