% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pathways.R
\name{pathway_chord}
\alias{pathway_chord}
\title{pathway_chord

Function used to plot a chord diagram of the ledge gene overlaps for pathways from fgsea results}
\usage{
pathway_chord(
  pathways,
  species_data = "human",
  outfile = "Pathways_Chord",
  save_plot = TRUE
)
}
\arguments{
\item{pathways}{a subset of distinct fsgea pathway results}

\item{species_data}{name of the species for which analysis applies, supports "human" or "mouse", defaults to "human"}

\item{outfile}{name of file for output}

\item{save_plot}{boolean of whether to save the plot as a file or just output to the console, defaults to TRUE and uses config information for file format}
}
\value{
image file of chord diagram pathway plot
}
\description{
pathway_chord

Function used to plot a chord diagram of the ledge gene overlaps for pathways from fgsea results
}
\examples{

pathway_chord(pathways, species_data = "human", outfile = paste0('GSEA_Pathways_Chord.', fileType), save_plot = TRUE)

}
