% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pathways.R
\name{pathway_intersect}
\alias{pathway_intersect}
\title{pathway_intersect

Function used to plot the ledge gene overlap for pathways from fgsea results}
\usage{
pathway_intersect(
  pathways,
  species_data = "human",
  outfile = "Pathways_Intersect",
  save_plot = TRUE
)
}
\arguments{
\item{pathways}{a subset of distinct fsgea pathway results}

\item{species_data}{name of the species for which analysis applies, supports "human" or "mouse", defaults to "human"}

\item{outfile}{name of file for output}

\item{save_plot}{boolean of whether to save the plot as a file or just output to the console, defaults to TRUE and uses config information for file format}
}
\value{
image file of pathway intersection plot

plotly interactive pathway intersection plot
}
\description{
pathway_intersect

Function used to plot the ledge gene overlap for pathways from fgsea results
}
\examples{

pathway_intersect(pathways, species_data = "human", outfile = paste0('Pathways_Intersect.', fileType), save_plot = TRUE)

}
