% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pathways.R
\name{pathway_overlap}
\alias{pathway_overlap}
\title{pathway_overlap

Function used to observe ledge gene overlap for pathways from fgsea results}
\usage{
pathway_overlap(pathways, all = TRUE, species_data = "human", for_plot = FALSE)
}
\arguments{
\item{pathways}{a subset of distinct fsgea pathway results}

\item{all}{boolean of whether or not to calculate pairwise overlaps for all ordered pathway pairs}

\item{species_data}{name of the species for which analysis applies, supports "human" or "mouse", defaults to "human"}
}
\value{
df containing the intersection of leading edge genes for selected pathways
}
\description{
Functions for pathway overlap and visualization
}
\examples{

pathways <- distinct(temp_res, pathway, .keep_all = TRUE)
pathways <- head(pathways[order(pathways$NES),], n = 10)
overlapOutput <- pathway_overlap(pathways, all = TRUE, species_data = "human")
}
