% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ROI_2D.R
\name{plot_ROI2D}
\alias{plot_ROI2D}
\title{ROI Plot}
\usage{
plot_ROI2D(
  data_df = NULL,
  annot_df = NULL,
  title = NULL,
  plot_type = "joy",
  skip_ROIs = TRUE,
  targets = "median",
  target_thresh = 0.5,
  target_quantile = TRUE,
  plot_clusters = TRUE,
  cluster_pal = "Set1",
  cluster_method = "complete",
  cluster_dist = "pearson",
  cluster_n = 6,
  cluster_min = 20,
  ROI_ID = "ROI_ID",
  AOI_ID = "Sample_ID",
  segment_ann = "Segment",
  segment_names = c("Tumor", "Stroma"),
  segment_colors = c("green3", "magenta3"),
  transf = NULL,
  cluster_transf = NULL,
  scale = TRUE,
  shape = "circle",
  hole_diameter = 5,
  width = 1,
  draw_bgd = TRUE,
  split_groups = TRUE,
  bgd_color = "white",
  bgd_fill = alpha("black", 0.5),
  joy_scale = 5,
  joy_alpha = 0.6,
  joy_lwd = 0.5,
  heat_theme = "Dark",
  heat_alpha = c(0, 0.9),
  heat_color = alpha("white", 0.5),
  ...
)
}
\arguments{
\item{data_df}{a data frame of targets / pathways / signatures scores / cell types vs AOIs to be plotted}

\item{annot_df}{a data frame of AOI annotations necessary to plot the data}

\item{title}{title to be added after ROI ID to each plot}

\item{plot_type}{type of plot, values include 'joy', 'bar', 'heatmap'}

\item{skip_ROIs}{convenience to just plot legend & grab gene sets}

\item{targets}{subset list or logic, either vector of booleans, character vector, or function for subseting including 'CV', 'IQR', 'sd', and other standard functions}

\item{target_thresh}{if a function is provided, use this threshold value to pick samples with value > thresh}

\item{target_quantile}{whether the target_thresh is a value or a quantile to filter to}

\item{plot_clusters}{whether clusters should be plotted in the static graph as well}

\item{cluster_pal}{cluster palette to use when plotting}

\item{cluster_method}{clustering type, 'average', 'complete', or 'ward' recommended}

\item{cluster_dist}{clustering distance, 'pearson', 'spearman', or values used by dist}

\item{cluster_n}{number of clusters to idenitfy}

\item{cluster_min}{the minimum size of a cluster for it to be considered distinct, otherwise breaks will be plotted with a neighboring cluster}

\item{ROI_ID}{name of ROI column or columns, list if multiple}

\item{AOI_ID}{name of AOI column}

\item{segment_ann}{segment annotation column}

\item{segment_names}{labels for segment, must be labels within the segment_ann column}

\item{segment_colors}{colors for segments used in joy plots and heatmap, not bar}

\item{transf}{'log2','logit','fraction','scale'}

\item{cluster_transf}{'log2','logit','fraction','scale'}

\item{shape}{type of plot to make based on original AOI shape - 'circle' or 'rect', rectangular is still in development}

\item{hole_diameter}{relative center diameter for window into ROI}

\item{width}{relative plotting area width}

\item{draw_bgd}{whether to draw the background behind plot if 'joy' or 'bar' used}

\item{split_groups}{whether to add a buffer between clusters and draw backgrounds seperately}

\item{bgd_color}{color for the border around backgrounds}

\item{bgd_fill}{color for the background fill}

\item{joy_scale}{scaling factor for joy plot}

\item{joy_alpha}{alpha for the fill of the joy plot}

\item{joy_lwd}{lwd for the joyplot}

\item{heat_theme}{Dark' = blacks with alpha, 'Light' = whites with alpha}

\item{heat_alpha}{range of alpha to used with the heatmap. 0 is pure color and 1 is a total overlay of the color}

\item{heat_color}{color for the border around the heatmap}
}
\value{
a list of ROI plots and legends to be subsequently saved to pdf or plotted with ggplotly
}
\description{
Takes data matrix, including expression, signatures, pathways and segment annotations, and creates graphs based on the ROI ID
}
