% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_plots.R
\name{qc_plots}
\alias{qc_plots}
\title{QC_plots

Creates QC plots of normalization factors and distributions of counts}
\usage{
qc_plots(
  data_matrix,
  norm_method = NULL,
  grp_var = NULL,
  ctrl_var = NULL,
  HKs = NULL,
  outdir,
  colors,
  txt_type = TRUE
)
}
\arguments{
\item{data_matrix}{the data matrix list}

\item{norm_method}{the selected norm method for output data matrix}

\item{grp_var}{the selected column name for the grouping analysis}

\item{ctrl_var}{the selected column name to control by for DE}

\item{HKs}{the list of housekeeping genes}

\item{outdir}{the output directory for plots}

\item{colors}{the list of specified colors}

\item{txt_type}{boolean specifying if txt files from Azorius were the input}
}
\value{
norm_counts chosen normalization method
}
\description{
QC_plots

Creates QC plots of normalization factors and distributions of counts
}
\examples{
 qc_plots(data_matrix = dfs, norm_method = "Q3", outdir = od)


}
