% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dataset.R
\name{read_dataset}
\alias{read_dataset}
\title{Reads in DSP-DA excel workbook for analysis and visualiation}
\usage{
read_dataset(xl_path, raw_xl, norm_xl, norm_type)
}
\arguments{
\item{xl_path}{string specifying path to dsp dataset excel workbooks}

\item{raw_xl}{string specifying name of raw collapsed count workbook}

\item{norm_xl}{string specifying name of normalized count workbook}

\item{norm_type}{string indicating type of normalization used in DSP-DA.
Accepted values: "Q3", "Neg", "HK"}
}
\value{
list of dataframes, one for each tab
}
\description{
Reads in DSP-DA excel workbook for analysis and visualiation
}
\examples{
read_dataset(xl_path="./testData/DA_data/Q3_normalization",
               raw_xl="ProbeQC - Default.xlsx",
               norm_xl="Q3 Norm.xlsx",
               norm_type="Q3")
}
