% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dataset.R
\name{reformat_dataset}
\alias{reformat_dataset}
\title{function to reformat DSP-DA data to be compatible with Hydra}
\usage{
reformat_dataset(curr_dfs, norm_type)
}
\arguments{
\item{curr_dfs}{list of current data frames resulting
from \code{\link{read_dataset}}}

\item{norm_type}{string identifying the type of normalization used}
}
\value{
list of data frames containing reformatted /code{curr_dfs}
}
\description{
function to reformat DSP-DA data to be compatible with Hydra
}
\examples{
Q3_dfs <- read_dataset(xl_path="./testData/DA_data/Q3_normalization",
                         raw_xl="ProbeQC - Default.xlsx",
                         norm_xl="Q3 Norm.xlsx",
                         norm_type="Q3")
reformat_dataset(curr_dfs=Q3_dfs, norm_type="Q3")
}
