% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ROC.R
\name{ROC_plot}
\alias{ROC_plot}
\title{Plot ROC Curve}
\usage{
ROC_plot(df, grouping_var, family, gene_group, outdir, genes)
}
\arguments{
\item{df}{negative normalized DSP count data transposed to wide format, one column per gene}

\item{grouping_var}{column from annotations to group by}

\item{family}{glm distribution, aka poisson}

\item{gene_group}{gene group of interest}

\item{outdir}{path to output directory}

\item{genes}{list of genes of interest}
}
\value{
image file of top 5 genes ROC curve

dataframe of AUC for each gene
}
\description{
Transposes counts df and creates ROC curve plot with top 5 genes with largest coefficients
}
\examples{
 ROC_plot(top5_coeff, gene_group, outdir)

}
