% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cell_decon.R
\name{run_cell_decon}
\alias{run_cell_decon}
\title{run_cell_decon

Runs our internal immune cell deconvolution algorithm, "InSituSort"}
\usage{
run_cell_decon(
  df = dfs,
  matrix = NULL,
  norm_method = norm_method,
  outdir = cellType_dir,
  tumor_high_ids = NULL,
  x = NULL,
  y = NULL,
  xlab = "",
  ylab = "",
  segment = segmentation,
  skip = NULL
)
}
\arguments{
\item{df}{dfs data frame containing azorius data}

\item{norm_method}{Raw data matric}

\item{outdir}{Directory to output figures}

\item{tumor_high_ids}{Names of AOIs with nearly 100% tumor content. Must be elements of the colnames of norm_counts.
If NULL, the decon model will ignore tumo-intrinsic expression, which usually works just fine.}

\item{x}{Horizontal coordinates on which to plot decon results  (e.g. from t-SNE) (optional)}

\item{y}{Vertical coordinates on which to plot decon results (e.g. from t-SNE) (optional)}

\item{xlab}{Name of the x-dimension on which results are plotted}

\item{ylab}{Name of the y-dimension on which results are plotted}

\item{segment}{Column with segmentation information}

\item{TME}{label for TME segmentations}
}
\value{
clustering information
}
\description{
run_cell_decon

Runs our internal immune cell deconvolution algorithm, "InSituSort"
}
\examples{

deconresults <- run_cell_decon(df = dfs, matrix = "Human_Cell_Landscape", norm_method = "Q3",
                               outdir = cellType_dir, tumor_high_ids = NULL, x = c(1,2,3,4),
                               y = c(4,3,2,1), xlab = "X coordinates", ylab = "Y coordinates",
                               segment = "aoi_type", skip = "Tumor")
}
