% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_analysis.R
\name{run_DE}
\alias{run_DE}
\title{DE Analysis}
\usage{
run_DE(
  df,
  control_var,
  grouping_var,
  base_level,
  residuals_dir,
  qq_dir,
  pval_cutoff,
  make_plots = TRUE,
  random_slope = TRUE
)
}
\arguments{
\item{df}{negative normalized DSP count data transposed to wide format, one column per gene}

\item{control_var}{column from annotations to use as experimental control variable}

\item{grouping_var}{column from annotations to test during DE analysis}

\item{base_level}{variable to use as base level for testing, all other vars looped through during analysis}

\item{residuals_dir}{path to output directory image subfolder for residual images}

\item{qq_dir}{path to output directory image subfolder for qq plots}

\item{pval_cutoff}{pvalue cutoff for significance from config file}

\item{make_plots}{logical as to whether qq & residual plots should be written}

\item{random_slope}{logical to turn on or off random slopes depending on study}
}
\value{
PNG file of top 5 genes ROC curve

dataframe of AUC for each gene
}
\description{
Runs a linear mixed effects model on the dataset to analyze differential expression using the LME4 package
}
\examples{
 run_DE(df, control_var, base_level, genes,residuals_dir, qq_dir, pval_cutoff)

}
