% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_GSEA.R
\name{run_GSEA}
\alias{run_GSEA}
\title{run_GSEA

Pathway Analysis helper function to run GSEA}
\usage{
run_GSEA(
  geneList = geneList,
  pathways = pathways,
  test = test,
  outdir = outdir
)
}
\arguments{
\item{geneList}{a list of ranked genes, the way that they are ranked is determined in run_pathways.R}

\item{pathways}{are the genesets that we are working with, default is Reactome}

\item{test}{is the test or contrast that is being compared for pathway analysis, from DE results}

\item{outdir}{the output directory for plots}
}
\value{
fgseaRes The results from GSEA analysis
}
\description{
run_GSEA

Pathway Analysis helper function to run GSEA
}
\examples{
 run_GSEA(geneList = geneList, pathways = pathways, test = test, outdir = outdir)

}
