% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_heatmap.R
\name{run_heatmap}
\alias{run_heatmap}
\title{run_heatmap

Generic call to heatmap functionality}
\usage{
run_heatmap(
  data = NULL,
  data_TargetID = "Gene",
  annot = NULL,
  samp_vars = NULL,
  fav_genes = NULL,
  ann_colors = NULL,
  logData = TRUE,
  targets = NULL,
  name = NULL,
  scale = "row",
  outdir = ".",
  fileType = "tiff",
  sortBy = NULL,
  sortBaseline = NULL,
  ...
)
}
\arguments{
\item{data}{the data to plot, assumed targets * samples}

\item{annot}{sample annotations}

\item{samp_vars}{annotations to include on heatmap}

\item{fav_genes}{favorite genes to highlight}

\item{ann_colors}{list of lists of colors}

\item{logData}{T/F - log2 transform data}

\item{targets}{used to subset dataset to only genes of interest or NULL for all}

\item{name}{title of graph and file}

\item{scale}{passed to pheatmap}

\item{outdir}{where to print file}

\item{fileType}{function to call for file printing}

\item{sortBy}{annotation variable to use for sorting the heatmap}

\item{width}{file width}

\item{height}{file height}
}
\value{

}
\description{
run_heatmap

Generic call to heatmap functionality
}
\examples{


}
