% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pathways.R
\name{run_pathways}
\alias{run_pathways}
\title{run_pathways

Pathway Analysis giving either coverage or GSEA}
\usage{
run_pathways(
  df,
  de_results,
  geneListrank,
  path_fc,
  path_pval,
  enrichment,
  outdir,
  custom_pathways_path = NULL,
  exclude_reactome = FALSE,
  ...
)
}
\arguments{
\item{df}{list of data frames containing input data}

\item{de_results}{the differential expression data}

\item{path_fc}{the fold change cutoff to use for gene list}

\item{path_pval}{the pvalue to threshold the enrichment results by}

\item{enrichment}{the type of enrichment to be returned}

\item{outdir}{the output directory for plots}

\item{custom_pathways_path}{the directory where custom gmt files are located}

\item{exclude_reactome}{to exclude reactome pathways from custom pathway analysis}

\item{genListrank}{the way to rank the gene list either by foldchange, pval or FDR}
}
\value{
enrich_res The results from your chosen enrichment
}
\description{
run_pathways

Pathway Analysis giving either coverage or GSEA
}
\examples{
 run_pathways(df = dfs, de_data = de_results, geneListrank=geneListrank, path_fc=path_fc, path_pval=path_pval, enrichment=enrichment, outdir = path_dir, custom_pathways_path = NULL, exclude_reactome = FALSE)

}
