% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_Q3.R
\name{run_Q3}
\alias{run_Q3}
\title{run_Q3

Run Q3 normalization}
\usage{
run_Q3(dataset = dfs, the_prob = 0.75)
}
\arguments{
\item{dataset}{the full dataset}

\item{the_prob}{a value between 0 and 1 for the quantile, default is 0.75}
}
\value{
dfs the new data object post q3 normalization, updated norm counts and properties with new q3 norm factor
}
\description{
run_Q3

Run Q3 normalization
}
\examples{
 run_Q3(dataset = dfs)

}
