% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_seqQC.R
\name{run_seqQC}
\alias{run_seqQC}
\title{DSP-NGS Sequencing QC}
\usage{
run_seqQC(
  dataframelist,
  loq = 2.5,
  qc_dir = "./",
  facet_column = "",
  fileType = "pdf",
  return_plots = FALSE
)
}
\arguments{
\item{dataframelist}{List of all 5 dataframes}

\item{loq}{Number of standard deviations desired for loq. Column must exist as GeoLOQ|SD|_|Pool| for each pool in the dataset, ex. GeoLOQ2.5_01. Default = 2.5}

\item{qc_dir}{Directory to output plot + QC table. Default = './'}

\item{facet_column}{Column to facet the line plot on if desired. Default = No facet.}

\item{fileType}{Type of file for ggsave to generate for the QC Line plot.}

\item{return_plots}{Boolean. If TRUE, a list of QC plots will be returned.}
}
\value{
Write QC metrics to file in QC directory

Write QC plots to file QC directory (QC of reads, QC of sequencing saturation)

QC plots as list if return_plots = TRUE.
      list('reads' = QC of reads, 'saturation' = QC of sequencing saturation)
}
\description{
Generates sequecing QC metrics for provided dataset
}
\examples{
   run_seqQC(dataframelist=dfs, loq=2.5, fileType='pdf')
}
