% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ssGSEA.R
\name{run_ssGSEA}
\alias{run_ssGSEA}
\title{run_ssGSEA

Pathway Analysis helper function to run ssGSEA}
\usage{
run_ssGSEA(
  norm_data = norm_data,
  annots,
  pathways = pathways,
  test = test,
  outdir = outdir
)
}
\arguments{
\item{norm_data}{the normalized log2 transformed data matrix}

\item{annots}{the annotations data.frame}

\item{pathways}{are the genesets that we are working with, default is Reactome}

\item{test}{is the test or contrast that is being compared for pathway analysis, from DE results}

\item{outdir}{the output directory for plots}
}
\value{
ssGSEA_results The results from ssGSEA analysis
}
\description{
run_ssGSEA

Pathway Analysis helper function to run ssGSEA
}
\examples{
 run_ssGSEA(norm_data = norm_data, pathways = pathways, test = test, outdir = outdir)

}
