% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_umap.R
\name{run_umap}
\alias{run_umap}
\title{run_umap

Runs t-Distributed Stochastic Neighbor Embedding (tsne) on dsp data}
\usage{
run_umap(
  df,
  pca = NULL,
  outdir,
  color = "cluster",
  symbol = "dsp_slide",
  size = TRUE,
  colors = colors
)
}
\arguments{
\item{df}{read in dsp excel workbook dataset}

\item{outdir}{output directory for figures}

\item{color}{category to color the graph by}

\item{symbol}{category to change symbols in the graph}

\item{size}{should the size of the points change by gene_count}

\item{outlier_cutoff}{multiplier for outlier cutoff}
}
\value{
none
}
\description{
run_umap

Runs t-Distributed Stochastic Neighbor Embedding (tsne) on dsp data
}
\examples{
run_umap(df, "file/path", outlier_cutoff, color = "slide_id", symbol = "TIS", size = TRUE)

}
