% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsp_de_analysis_parallel.R
\name{send_to_log}
\alias{send_to_log}
\title{Internal helper function for writing to log.}
\usage{
send_to_log(grouping_var, base_level, control_var)
}
\description{
Internal helper function for writing to log.
}
